Create scorecard-service and makes all tests pass
Refactor get-scorecard and get-user-scorecards and ensure that it works the same
