+++ api/lib/domain/usecases/get-scorecard.js
@@ -1,8 +1,6 @@
-const _ = require('lodash');
 const { UserNotAuthorizedToAccessEntity } = require('../errors');
-const Scorecard = require('../models/Scorecard');
 
-module.exports = async ({ authenticatedUserId, scorecardId, knowledgeElementRepository, competenceRepository, competenceEvaluationRepository }) => {
+module.exports = async ({ authenticatedUserId, scorecardId, knowledgeElementRepository, competenceRepository, competenceEvaluationRepository, scorecardService }) => {
 
   const [userId, competenceId] = scorecardId.split('_');
   if (parseInt(authenticatedUserId) !== parseInt(userId)) {
@@ -15,32 +13,6 @@ module.exports = async ({ authenticatedUserId, scorecardId, knowledgeElementRepo
     competenceEvaluationRepository.findByUserId(authenticatedUserId),
   ]);
 
-  const sortedKEGroupedByCompetence = _.groupBy(userKEList, 'competenceId');
-  const knowledgeElementsOfCompetence = sortedKEGroupedByCompetence[competence.id];
-  const totalEarnedPixByCompetence = _.sumBy(knowledgeElementsOfCompetence, 'earnedPix');
-
-  return new Scorecard({
-    id: scorecardId,
-    name: competence.name,
-    description: competence.description,
-    competenceId: competence.id,
-    index: competence.index,
-    area: competence.area,
-    earnedPix: totalEarnedPixByCompetence,
-    status: _getStatus(knowledgeElementsOfCompetence, competence.id, competenceEvaluations)
-  });
+  return scorecardService.createScorecard(authenticatedUserId, userKEList, competence, competenceEvaluations);
 };
 
-function _getStatus(knowledgeElements, competenceId, competenceEvaluation) {
-  if (_.isEmpty(knowledgeElements)) {
-    return 'NOT_STARTED';
-  }
-
-  const competenceEvaluationForCompetence = _.find(competenceEvaluation, { competenceId });
-  const stateOfAssessment = _.get(competenceEvaluationForCompetence, 'assessment.state');
-  if (stateOfAssessment === 'completed') {
-    return 'COMPLETED';
-  }
-  return 'STARTED';
-
-}
